+++ CHANGELOG.md
@@ -10,6 +10,8 @@ Released at: not yet released
 
 2. Convert special chars to their HTML equivalents.
 
+3. Remove bang ('!') token from all routes.
+
 ## 1.5.1
 
 Released at: 18.03.2020